% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix Function}
\usage{
cor_matrix(x)
}
\arguments{
\item{years}{extend from 1950 to 2017}
}
\description{
This function allows you to return a correlation matrix for each year in the nba dataset
}
\examples{
cor_matrix(1950) would yield a 22x22 matrix
}
\keyword{correlation}
\keyword{matrix}
\keyword{year,}
