% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_points.R
\name{highest_points}
\alias{highest_points}
\title{Highest Points Function}
\usage{
highest_points(x)
}
\arguments{
\item{years}{extend from 1950 to 2017}
}
\description{
This function allows you to find the highest points for each year in the nba dataset
}
\examples{
highest_points(1950) would yield 1865
}
\keyword{points}
\keyword{year,}
