% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(x)
}
\arguments{
\item{years}{extend from 1950 to 2017}
}
\description{
This function allows you to find the oldest player for each year in the nba dataset
}
\examples{
oldest_player(1950) would yield 36
}
\keyword{age}
\keyword{year,}
